Split ahp-handlers.ts into focused modules (#1472)
Decompose the 1438-line monolithic handler file into 7 single-concern
modules while preserving the public API (mountAhpServer + MountAhpServerOptions).

New modules:
- ahp-types.ts: shared interfaces (ForwarderState, ResourceWatchEntry, ClientState)
- channel-codec.ts: URI encoding/decoding and protocol constants
- forwarder.ts: event forwarding loop, mapper integration, orphan/status rescue
- resource-watch.ts: filesystem watch lifecycle (create, start, stop, flush)
- handlers/session-handlers.ts: create, dispose, list, initialize
- handlers/subscribe-handlers.ts: session + resource-watch subscription routing
- handlers/action-handlers.ts: dispatch, authenticate

ahp-handlers.ts retains only the routing switch, onDisconnect orchestrator,
and JSON-RPC response helpers (~240 lines).
